# docs/swagger.yaml
basePath: /
definitions:
  handlers.NewPassword:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  task.Task:
    properties:
      is_completed:
        default: false
        type: boolean
      parent_task_id:
        type: integer
      points:
        type: integer
      user_id:
        type: integer
    required:
    - user_id
    type: object
  user.User:
    properties:
      is_admin:
        default: false
        type: boolean
      password:
        type: string
      rating:
        type: integer
      username:
        type: string
    required:
    - password
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  license:
    name: Apache 2.0
  title: Project Swagger
  version: "0.1"
paths:
  /admin/changeAdminPassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: new password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.NewPassword'
      produces:
      - application/json
      responses: {}
      security:
      - BasicAuth: []
      summary: ChangeAdminPassword
      tags:
      - Admin
  /admin/createTask:
    post:
      consumes:
      - application/json
      parameters:
      - description: task data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/task.Task'
      produces:
      - application/json
      responses: {}
      security:
      - BasicAuth: []
      summary: CreateTask
      tags:
      - Admin
  /admin/createUser:
    post:
      consumes:
      - application/json
      parameters:
      - description: user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses: {}
      security:
      - BasicAuth: []
      summary: CreateUser
      tags:
      - Admin
  /user/completeTask/{taskId}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      responses: {}
      security:
      - BasicAuth: []
      summary: CompleteTask
      tags:
      - User
  /user/getRating:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - BasicAuth: []
      summary: GetRating
      tags:
      - User
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"